% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_codes.R
\name{get_country_list}
\alias{get_country_list}
\title{Get a list of the countries that can be queried}
\usage{
get_country_list()
}
\value{
A list of country codes
}
\description{
get a list of the country codes that you can query from the API
}
\examples{
require(SDGsR)
cnts<-get_country_list()[228]
}
