% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{get_indicator_data}
\alias{get_indicator_data}
\title{get_indicator_data}
\usage{
get_indicator_data(indicator = "15.6.1")
}
\arguments{
\item{indicator}{the code for the indicator you are interested in}
}
\value{
dataframe of indicator and values
}
\description{
get indicator data for a specific indicator
}
\examples{
get_indicator_data (indicator="15.6.1")
}
