% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_example_data.R
\name{get_indicator_example_data}
\alias{get_indicator_example_data}
\title{Get and example of indicator data for 15.6.1, 15.9.1 and 15.a.1}
\usage{
get_indicator_example_data()
}
\value{
tibble of indicator data
}
\description{
get the example data for the indicator example shiny
}
\examples{
\dontrun{
require(SDGsR)
get_indicator_example_data()
}
}
\keyword{Internal}
