% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_list.R
\name{get_indicator_list}
\alias{get_indicator_list}
\title{get_indicator_list}
\usage{
get_indicator_list()
}
\value{
A dataframe of indicators and the description
}
\description{
get a list of the indicators for each goal and target
}
\examples{
get_indicator_list()
}
