% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SDGS_goals.R
\name{get_SDGs_goals}
\alias{get_SDGs_goals}
\title{get_SDGS_goals}
\usage{
get_SDGs_goals()
}
\value{
writes an .RDS file to a "data" folder listing Goals, Targets and indicators
}
\description{
calls the UN SDGs API to get the Goals data
}
\examples{
get_SDGs_goals()
}
