% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SDGS_goals_titles.R
\name{get_SDGs_goals_titles}
\alias{get_SDGs_goals_titles}
\title{Get the titles of the SDGS goals}
\usage{
get_SDGs_goals_titles()
}
\value{
dataframe of goal titles
}
\description{
calls the UN SDGs API to get the titles for each Goal
}
