% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_codes.R
\name{lookup_country}
\alias{lookup_country}
\title{lookup_country_codes}
\usage{
lookup_country(code = "M49", country)
}
\arguments{
\item{code}{M49 or ISO3 alpha code}

\item{country}{the name of country (use get_country_list to see the accepted names of countries)}
}
\value{
A M49 or ISO3 alphacode
}
\description{
Lookup a country code for a country
}
\examples{
lookup_country(code="M49", country="Norway")
}
