% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_country_codes}
\alias{lookup_country_codes}
\title{Lookup country codes}
\format{
A data frame with 249 rows and 3 variables:
\describe{
\item{\code{country_or_area}}{character COLUMN_DESCRIPTION}
\item{\code{m49_code}}{double COLUMN_DESCRIPTION}
\item{\code{iso_alpha3_code}}{character COLUMN_DESCRIPTION}
}
}
\usage{
lookup_country_codes
}
\description{
A dataframe used to lookup
country codes in the M49 and ISO_alpha3 format
}
\keyword{datasets}
