% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGs_colours.R
\name{SDGs_cols}
\alias{SDGs_cols}
\title{Get the SDGS colour pallette}
\usage{
SDGs_cols(...)
}
\arguments{
\item{...}{arguments for passing to other functions}
}
\value{
hex colour for each Goal
}
\description{
get the SDGS colour for each Goal
}
\examples{
\dontrun{
Norway <- SDGsR::get_indicator(Country = "578", indicator = "15.4.1")
Norway \%>\%
  dplyr::select(timePeriodStart, value, seriesDescription) \%>\%
  ggplot2::ggplot(aes(timePeriodStart, value)) +
  ggplot2::geom_point(colour = SDGs_cols("Goal15")) +
  ggplot2::ggtitle(label = paste0(Norway$seriesDescription[1]))
  }
}
