% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{get_indicator}
\alias{get_indicator}
\title{get_indicator}
\usage{
get_indicator(Country, indicator)
}
\arguments{
\item{Country}{the M49 code for the country you are interested in}

\item{indicator}{the code for the indicator you are interested in}
}
\value{
dataframe of indicator and values
}
\description{
get indicators for a specific country and indicator
}
\examples{
get_indicator(Country="12", indicator="15.6.1")
}
