% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_example_data.R
\name{get_indicator_example_data}
\alias{get_indicator_example_data}
\title{get_indicator_example_data}
\usage{
get_indicator_example_data()
}
\value{
tibble of indicator data
}
\description{
get the example data for the indicator example shiny
}
\examples{
\dontrun{
get_indicator_example_data()
}
}
