% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SDGS_goals.R
\name{get_SDGs_goals}
\alias{get_SDGs_goals}
\title{Get the SDGs Goals, Targets and indicators as a dataframe}
\usage{
get_SDGs_goals(pathtosave = NULL)
}
\arguments{
\item{pathtosave}{An optional path to save the data as an RDS}
}
\value{
dataframe listing Goals, Targets and indicators
}
\description{
calls the UN SDGs API to get the Goals data
}
\examples{
get_SDGs_goals()
}
