% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample(example)
}
\arguments{
\item{example}{Can only be "indicator.R" at present}
}
\value{
A shiny app with plotly interactive map
}
\description{
Run ShinyApp examples
}
